articles/microservices/article_generation/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues


azure-eventhub
newsapi==0.1.1
newsapi-python==0.2.6
pandas==1.4.0
numpy==1.22.0
requests==2.27.1
bs4
